Refactor: make it natural
